issue82 Clean up TestRoomService

+++ Museum-Backend/src/main/java/ca/mcgill/ecse321/museum/service/RoomService.java
@@ -185,6 +185,12 @@ public class RoomService {
       throw new IllegalArgumentException("Room does not exist");
     }
 
+    // Check if there are artworks in the room
+    List<Artwork> artworks = toList(artworkRepository.findArtworkByRoom(room));
+    if (artworks.size() > 0) {
+      throw new IllegalArgumentException("There are still artworks in the room, remove them first");
+    }
+
     // Delete room
     roomRepository.delete(room);
     return room;
@@ -203,12 +209,6 @@ public class RoomService {
     // Error handling
     if (roomType == null) {
       throw new IllegalArgumentException("Room type cannot be empty");
-    }
-
-    // Check if there are artworks in the room
-    List<Artwork> artworks = toList(artworkRepository.findArtworkByRoom(room));
-    if (artworks.size() > 0) {
-      throw new IllegalArgumentException("There are still artworks in the room, remove them first");
     }
 
     // Get maximum number of artworks
